content/media/test/test_bug686942.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=686942
-->

<head>
  <title>Test for Bug 448534</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=686942">Mozilla Bug 686942</a>
<pre id="test">
<script class="testbody" type="text/javascript">

if (navigator.platform.startsWith("Win")) {
  SimpleTest.expectAssertions(0, 2);
}

var manager = new MediaTestManager;

function onloaded(event) {
  var v = event.target;
  v.currentTime = v.duration;
  return;
}

function checkNotPlaying(v) {
  ok(v.currentTime == 0, "Should not be playing after seek to end and back to beginning");
  v._finished = true;
  manager.finished(v.token);
}

function onseeked(event) {
  var v = event.target;
  setTimeout(function() { checkNotPlaying(v); }, 500);
}

function onended(event) {
  var v = event.target;
  if (v._finished)
    return;
  v.addEventListener("seeked", onseeked, false);
  v.currentTime = 0;
}

function startTest(test, token) {
  var v = document.createElement('video');
  v.token = token;
  manager.started(token);
  v.src = test.name;
  v._played = false;
  v._finished = false;
  v.addEventListener("loadedmetadata", onloaded, false);
  v.addEventListener("ended", onended, false);

  document.body.appendChild(v);
}

manager.runTests(gSmallTests, startTest);

</script>
</pre>
</body>
</html>
